+++ apps/shared/storage_function_banner_delegate.cpp
@@ -8,12 +8,12 @@ namespace Shared {
 
 void StorageFunctionBannerDelegate::reloadBannerViewForCursorOnFunction(CurveViewCursor * cursor, Ion::Storage::Record record, StorageFunctionStore * functionStore, char symbol) {
   ExpiringPointer<StorageFunction> function = functionStore->modelForRecord(record);
-  constexpr int bufferSize = k_maxNumberOfCharacters+PrintFloat::bufferSizeForFloatsWithPrecision(Constant::LargeNumberOfSignificantDigits);
+  constexpr size_t bufferSize = k_maxNumberOfCharacters+PrintFloat::bufferSizeForFloatsWithPrecision(Constant::LargeNumberOfSignificantDigits);
   char buffer[bufferSize];
   const char * space = " ";
   const char * legend = "0=";
   int legendLength = strlen(legend);
-  int numberOfChar = 0;
+  size_t numberOfChar = 0;
   strlcpy(buffer, legend, bufferSize);
   numberOfChar += legendLength;
   buffer[0] = symbol;
